+++ @navikt/core/react/src/overlays/overlay/hooks/useAnimationsFinished.ts
